DictionaryLuaFormatter: Fix precision bug

+++ src/misc/dictionaryluaformatter.cpp
@@ -65,7 +65,7 @@ std::string DictionaryLuaFormatter::formatValue(const Dictionary& dictionary,
     }
 
     if (dictionary.hasValue<glm::dvec4>(key)) {
-        glm::vec4 vec = dictionary.value<glm::dvec4>(key);
+        glm::dvec4 vec = dictionary.value<glm::dvec4>(key);
         return "{" + std::to_string(vec.x) + "," +
             std::to_string(vec.y) + "," +
             std::to_string(vec.z) + "," +
@@ -73,20 +73,20 @@ std::string DictionaryLuaFormatter::formatValue(const Dictionary& dictionary,
     }
 
     if (dictionary.hasValue<glm::dvec3>(key)) {
-        glm::vec3 vec = dictionary.value<glm::dvec3>(key);
+        glm::dvec3 vec = dictionary.value<glm::dvec3>(key);
         return "{" + std::to_string(vec.x) + "," +
             std::to_string(vec.y) + "," +
             std::to_string(vec.z) + "}";
     }
 
     if (dictionary.hasValue<glm::dvec2>(key)) {
-        glm::vec2 vec = dictionary.value<glm::dvec2>(key);
+        glm::dvec2 vec = dictionary.value<glm::dvec2>(key);
         return "{" + std::to_string(vec.x) + "," +
             std::to_string(vec.y) + "}";
     }
 
     if (dictionary.hasValue<double>(key)) {
-        float value = dictionary.value<double>(key);
+        double value = dictionary.value<double>(key);
         return std::to_string(value);
     }
